% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socnet.r
\docType{data}
\name{socnet-archives}
\alias{socnet-archives}
\alias{SOCNET_ARCHIVES_WWW}
\alias{LISTS_UFL_WWW}
\title{SOCNET's listserv web address}
\format{Character scalars with web addresses.}
\usage{
SOCNET_ARCHIVES_WWW

LISTS_UFL_WWW
}
\value{
Currently, the values of both variables is
\itemize{
\item \code{SOCNET_ARCHIVES_WWW}: \Sexpr{socnet::SOCNET_ARCHIVES_WWW}
\item \code{LISTS_UFL_WWW}: \Sexpr{socnet::LISTS_UFL_WWW}
}
}
\description{
SOCNET's listserv web address
}
\keyword{datasets}
