% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socnet.r
\name{socnet_list_subjects}
\alias{socnet_list_subjects}
\title{Retrieve the table with the links from a particular archive}
\usage{
socnet_list_subjects(url, cached = FALSE)
}
\arguments{
\item{url}{Character scalar. A web address from a particular SOCNET listserv}

\item{cached}{Logical scalar. When \code{TRUE}, the function will check whether
the website is available in \Sexpr{system.file("cache", package="socnet")}
and read that instead of downloading the current version of the website.}
}
\description{
Retrieve the table with the links from a particular archive
}
\details{
Using \code{cached=FALSE} makes more sense if the user is trying to fetch versions
of SOCNET that require update, for example, the "SOCNET archives" website
itself.
}
\seealso{
Other SOCNET webscrapping tools: \code{\link{socnet_list_archives}},
  \code{\link{socnet_parse_subject}}
}
\concept{SOCNET webscrapping tools}
