% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socnet.r
\name{socnet_parse_subject}
\alias{socnet_parse_subject}
\title{Retrieve contents from a particular subject}
\usage{
socnet_parse_subject(url)
}
\arguments{
\item{url}{Character scalar. A web address to a SOCNET Listserv sibject}
}
\description{
Retrieve contents from a particular subject
}
\examples{
\dontrun{
socnet_parse_subject(
  "https://lists.ufl.edu/cgi-bin/wa?A2=ind1711&L=SOCNET&T=0&F=&S=&P=47265"
)
}
}
\seealso{
Other SOCNET webscrapping tools: \code{\link{socnet_list_archives}},
  \code{\link{socnet_list_subjects}}
}
\concept{SOCNET webscrapping tools}
