% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{subjects}
\alias{subjects}
\title{Complete dataset of SOCNET emails}
\format{A data frame with 200 observations and the following variables}
\source{
\Sexpr{readLines(system.file("cache","readme.md", package="socnet"), warn=FALSE)}
}
\usage{
subjects
}
\description{
Complete dataset of SOCNET emails
}
\keyword{datasets}
